allow relative path in streamlocal forwarding
When forwarding a Unix-domain socket, the remote socket path must be
absolute (otherwise the forwarding fails later).  However, guessing
absolute path on the remote end is sometimes not straightforward,
because the file system location may vary for many reasons, including
the system installation, the choices of NFS mount points, or the
remote user ID.

To allow ssh clients to request remote socket forwarding without
knowledge of the remote system, this patch enables the use of relative
path in remote socket forwarding.  If a relative path is requested as
remote_socket, it is expanded from StreamLocalBindRootDirectory, a new
option added to sshd_config.

This feature would be particularly useful if the remote system is
capable of user runtime directory, as managed by pam_systemd:
https://www.freedesktop.org/software/systemd/man/pam_systemd.html

The applications locating sockets under the runtime directory could
transparently access the forwarded sockets, with the following setting:

  StreamLocalBindRootDirectory /run/user/%i